Twitter_clone/Switter-backend/requirements.txt
flask
flask_smorest
flask_restful
marshmallow
flask_jwt_extended
flask_sqlalchemy
sqlalchemy
alembic
passlib
pipenv
python-dotenv
flask_migrate
requests